chore: Move taken logic to Taken class
