+++ src/app/zaken/Taken.ts
@@ -1,7 +1,33 @@
+import axios from 'axios';
 import { OpenZaakClient } from './OpenZaakClient';
 
 export class Taken {
+
+  static takenFromSecret(secret: string) {
+    if (process.env.USE_TAKEN !== 'true') {
+      return false;
+    }
+    if (!process.env.VIP_TOKEN_BASE_URL) {
+      throw Error('No VIP_TOKEN_BASE_URL provided');
+    }
+    const instance = axios.create(
+      {
+        baseURL: process.env.VIP_TOKEN_BASE_URL,
+        headers: {
+          Authorization: 'Token ' + secret,
+        },
+      },
+    );
+    const openZaakClient = new OpenZaakClient({
+      baseUrl: new URL(process.env.VIP_TOKEN_BASE_URL),
+      axiosInstance: instance,
+    });
+
+    return new Taken(openZaakClient);
+  }
+
   private client: OpenZaakClient;
+
   constructor(client: OpenZaakClient) {
     this.client = client;
   }
@@ -42,3 +68,5 @@ export class Taken {
     return date.toLocaleDateString('nl-NL', { day: 'numeric', month: 'long', year: 'numeric' });
   }
 }
+
+
